+++ src/main/java/wooteco/subway/exception/notfound/DataNotFoundException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.notfound;
 
 public class DataNotFoundException extends RuntimeException {
