+++ src/main/java/wooteco/subway/exception/notfound/LineNotFoundException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.exception.notfound;
+
+public class LineNotFoundException extends DataNotFoundException {
+
+    public LineNotFoundException(String message) {
+        super(message);
+    }
+    
+}
