+++ src/main/java/wooteco/subway/exception/notfound/SectionNotFoundException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.exception.notfound;
+
+public class SectionNotFoundException extends DataNotFoundException {
+
+    public SectionNotFoundException(String message) {
+        super(message);
+    }
+    
+}
