+++ src/main/java/wooteco/subway/exception/notfound/StationNotFoundException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.exception.notfound;
+
+public class StationNotFoundException extends DataNotFoundException {
+
+    public StationNotFoundException(String message) {
+        super(message);
+    }
+    
+}
