+++ src/test/java/wooteco/subway/domain/SectionsTest.java
@@ -8,9 +8,9 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wooteco.subway.exception.DataNotFoundException;
 import wooteco.subway.exception.duplicate.DuplicateSectionException;
 import wooteco.subway.exception.invalidrequest.InvalidSectionCreateRequestException;
+import wooteco.subway.exception.notfound.DataNotFoundException;
 
 class SectionsTest {
